Finished Week 3. Kinda.
It's not quite as polished as I want, but dang it, it works! I spent a lot of time figuring this out while also constantly feeling burnt out, so this is the best I've got. Still need to add CRUD for courses and update CRUD for grades, but that's the next lesson... that I'll have to do sometime. Oh no.